+++ README.md
@@ -178,7 +178,7 @@ You can also get information about how to use the plugin from code in [`examples
 
 It is a list of handlers to scroll the element.
 
-**Default**: `['click-rail', 'drag-scrollbar', 'keyboard', 'wheel', 'touch']`
+**Default**: `['click-rail', 'drag-thumb, 'keyboard', 'wheel', 'touch']`
 
 ### `wheelSpeed {Number}`
