+++ src/handlers/drag-thumb.js
